tidy files
